Make YAML features optional
