+++ apispec/ext/flask.py
@@ -75,7 +75,7 @@ from flask import current_app
 from flask.views import MethodView
 
 from apispec.compat import iteritems
-from apispec import BasePlugin, utils
+from apispec import BasePlugin, yaml_utils
 from apispec.exceptions import APISpecError
 
 
@@ -111,13 +111,13 @@ class FlaskPlugin(BasePlugin):
     def path_helper(self, operations, view, **kwargs):
         """Path helper that allows passing a Flask view function."""
         rule = self._rule_for_view(view)
-        operations.update(utils.load_operations_from_docstring(view.__doc__) or {})
+        operations.update(yaml_utils.load_operations_from_docstring(view.__doc__) or {})
         if hasattr(view, 'view_class') and issubclass(view.view_class, MethodView):
             for method in view.methods:
                 if method in rule.methods:
                     method_name = method.lower()
                     method = getattr(view.view_class, method_name)
-                    docstring_yaml = utils.load_yaml_from_docstring(method.__doc__)
+                    docstring_yaml = yaml_utils.load_yaml_from_docstring(method.__doc__)
                     operations[method_name] = docstring_yaml or dict()
         path = self.flaskpath2openapi(rule.rule)
         app_root = current_app.config['APPLICATION_ROOT'] or '/'
