+++ dev-requirements.txt
@@ -1,5 +1,4 @@
 invoke==1.1.0
-pyyaml==3.13  # pyup: <4.0
 
 # Syntax checking
 flake8==3.5.0
@@ -9,6 +8,7 @@ Flask==1.0.2
 marshmallow==2.15.3
 tornado==5.1
 bottle==0.12.13
+pyyaml==3.13  # pyup: <4.0
 
 # testing
 pytest==3.6.3
